A new "is" function to partially match state
The standard state.match function requires the fully qualified state but this
machine has quite nested states and is tedious and error prone to type out.
It also makes refactoring the nesting of the states harder.

This new method just checks if the fully qualified state includes the given
string. While this may be error prone if states share parts of their names, I
think it'll be easier to keep the state names unique in this case.