Fix the sticky hover state issue
This sets the hover background to transparent if the device does not support
hovering and also sets the active state as the original hover state.

The result of this change is that on mobile devices, the hover state will be
active while the user has tapped the button, but then once they let go, the
hover state disappears.