Make disabled steps actually get skipped
There is a ton of repetition due to needing to filter out all the blocks that are
disabled everytime they're accessed.

Afaik, XState doesn't have the concept of computer context values but this
probably needs to be extracted into some sort of wrapper object with methods to
return the filtered list of blocks in one place.