Allow moving program, fix deselect bug, add loaded state
The logic to deselect an already selected item was in an XState action but did
not transition to the desired "program not selected" state.
This is now done with a new DESELECT_PROGRAM event.

The loaded state was needed so a program can be moved any time the data is
loaded, not just if a program has been selected.