Remove the emoji field
If users want this, they can add it to the name field.
This saves me from needing to implement a special emoji picker.

I also removed the strict zod check because old programs with the emoji key
would appear as invalid, and the UI doesn't explain why nor provide a way of
fixing it other than restoring the default data.