Use uncontrolled inputs and debounce some fields
While using uncontrolled inputs is not my usual preference, it seemed like the
easiest way to use the `use-debounce` library (judging by their docs).

I've used uncontrolled inputs and debounced send functions to prevent the typing
in the fields being blocked by the state machine updates, which seem pretty
slow!

It's likely slow due to writing to parsing the objects with Zod and writing
every change to local storage on key input.
The debounce allows the typing to feel instant, as because it's uncontrolled, it
does become out of sync while the update is pending during typing.

I could also keep these components controlled with local state and still
debounce the send, but I'm not sure if I'd gain anything and this was simpler.

I didn't use this technique for the duration fields so the problem is still
present there. As this is a custom component, it is a little trickier and users
generally only need to type 1 or 2 numbers in so hopefully it won't be too
noticeable - if it is, I'll update those too.