Basic voice recognition working for pause blocks!
This can be extended for "start", "stop", etc. later on.

It's currently using a deprecated "send" action created, which needs to be
replaced with "sendTo" but I'm not sure how to refer to the appropriate "actor"
so I need to read the docs more first!

The voice recognition code also waits for the recognised word to be "final",
which means there is a delay before an action is taken.
This can be improved by looking for the keywords in the "interim" results but
this needs a bit more work.