Make the voice recognition last forever with hacks
Also made the UI look a little nicer.

Currently, the voice commands shown to the user and what the machine actually
listens for are not coming from a single source of truth and has a risk of
getting out of sync - this could be improved in the future by making the UI
reflect what is actually being listened for.