Refresh the ConfigScreen when the program changes
There was a bug where the uncontrolled title and description components were
holding the state of the initially selected program and wouldn't update when a
new program is selected.

These controls should probably be controlled instead, but that causes a delay
for writing to state (possibly due to writing to local storage or the zod
validation).

This satisfies both requirements of updating the controls when the program is
changed and decoupling the input from writing to state.

Still seems like the uncontrolled inputs could get out of sync if updated by any
other means but as the UI doesn't allow that, hopefully it won't be a problem in
practice.