Make the header inset
This looks better when used as a PWA without the browser chrome on
Android, because the OS status bar becomes the same color as the
page background color, not the header background color.

This makes the page background color flow around the header and all the
way to the top of the phone seamlessly.